Add `measure-artifact-size` feature

+++ web/src/wasmJsMain/kotlin/nl/jacobras/codebaseobserver/ui/chart/ChartData.kt
@@ -2,5 +2,5 @@ package nl.jacobras.codebaseobserver.ui.chart
 
 data class ChartData(
     val xLabels: List<String>,
-    val yValues: List<Int>
+    val yValues: List<Long>
 )
